test: split slow nuts
